removed unnecessary code related to classic quadrature and added color output to sphinx build
