+++ docs/Makefile
@@ -5,7 +5,7 @@
 # https://jefflirion.github.io/sphinx-github-pages.html
 
 # You can set these variables from the command line.
-SPHINXOPTS    =
+SPHINXOPTS    = --color
 SPHINXBUILD   = sphinx-build
 SPHINXPROJ    = probnum
 SOURCEDIR     = source
