--- src/probnum/quad/bayesian/__init__.py
@@ -1,1 +1,0 @@
-from probnum.quad.bayesian.bayesquadrature import *
