+++ src/probnum/quad/bayesian_quadrature/__init__.py
@@ -1,0 +1,2 @@
+from ._bayesian_quadrature import *
+from .bayesquad import *
