--- src/probnum/quad/polynomial/__init__.py
@@ -1,2 +1,0 @@
-from probnum.quad.polynomial.clenshawcurtis import *
-from probnum.quad.polynomial.polynomialquadrature import *
