+++ src/probnum/utils/__init__.py
@@ -1,6 +1,5 @@
 from .argutils import *
 from .arrayutils import *
-from .fctutils import *
 from .randomutils import *
 
 # Public classes and functions. Order is reflected in documentation.
@@ -13,6 +12,5 @@ __all__ = [
     "as_shape",
     "assert_is_1d_ndarray",
     "assert_is_2d_ndarray",
-    "assert_evaluates_to_scalar",
     "derive_random_seed",
 ]
